Allow legacy Python types for `Table` schema construction, fix root `__all__`
