+++ rust/perspective-python/perspective/__init__.py
@@ -16,7 +16,6 @@ __all__ = [
     "Server",
     "Client",
     "PerspectiveError",
-    "PerspectiveWidget",
     "ProxySession",
 ]
